oose/requirements.txt
amqp==1.4.9
anyjson==0.3.3
appnope==0.1.0
autopep8==1.3.3
Babel==2.5.1
billiard==3.3.0.23
celery==3.1.21
certifi==2018.1.18
cffi==1.7.0
chardet==3.0.4
cloudinary==1.1.3
coreapi==2.3.3
coreschema==0.0.4

cymem==1.31.2
cytoolz==0.8.2
decorator==4.1.2
defusedxml==0.5.0
diff-match-patch==20121119
dill==0.2.7.1
django-appconf==1.0.1
django-braces==1.8.1
django-choices==1.4.2
django-cors-headers==2.1.0
django-crontab==0.7.1
django-debug-toolbar==1.4
django-extensions==1.9.1
django-fcm==0.0.2
django-filter==0.12.0
django-fsm==2.3.0
django-fsm-admin==1.2.2
django-fsm-log==1.3.0
django-grappelli==2.8.1
django-haystack==2.6.1
django-haystack-elasticsearch5==0.7
django-import-export==0.4.2
django-model-utils==3.1.1
django-money==0.7.6
django-oauth-toolkit==0.10.0
django-queryinspect==0.1.0
django-redis-cache==1.6.5
django-rest-framework-social-oauth2==1.0.4
django-rest-swagger==0.3.6
django-silk==1.0.0
djangorestframework==3.3.2
djangorestframework-bulk==0.2.1
djangorestframework-gis==0.10.1
djorm-ext-pgfulltext==0.10
drfdocs==0.0.11
elasticsearch==5.0.0

et-xmlfile==1.0.1
flower==0.9.1
ftfy==4.4.3
html5lib==1.0b8
idna==2.6
ipython==6.2.0
ipython-genutils==0.2.0
itypes==1.1.0
jdcal==1.3
jedi==0.10.2
Jinja2==2.10
kombu==3.0.34
Markdown==2.6.6
MarkupSafe==1.0
msgpack-numpy==0.4.1
msgpack-python==0.5.4
murmurhash==0.28.0
ndg-httpsclient==0.4.2
newrelic==2.74.0.54
oauthlib==1.0.3
odfpy==1.3.5
openpyxl==2.4.8
pathlib==1.0.1
pexpect==4.2.1
pickleshare==0.7.4
plac==0.9.6
preshed==1.0.0
prompt-toolkit==1.0.15

ptyprocess==0.5.2
py-moneyed==0.6.0
pyasn1==0.1.9
pycodestyle==2.3.1
pycparser==2.14
pycrypto==2.6.1
Pygments==2.2.0
PyJWT==1.5.3
pyOpenSSL==16.0.0
python-dateutil==2.5.0
python-firebase==1.2
python-social-auth==0.2.14
python3-openid==3.1.0
pytz==2015.7
PyYAML==3.11
raven==6.3.0
razorpay==0.2.0
redis==2.10.5
regex==2017.4.5
requests==2.18.4
requests-oauthlib==0.8.0
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.2.3
tablib==0.12.1
termcolor==1.1.0
thinc==6.10.2
toolz==0.9.0
tornado==4.2
tqdm==4.19.7
traitlets==4.3.2
ujson==1.35
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.37.0
wrapt==1.10.11
xlrd==1.1.0
xlwt==1.3.0
xmltodict==0.11.0
